pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["Wojciech Malarski <w.malarski@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.34.0"
grpcio-tools = "^1.34.0"
click = "^7.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
